Update job status & error monitoring mechanism
All job filenames and ids in Airflow DB are generated
on the base of BioWardrobe uid. This allows to catch
the cases when new job file should be generated,
but some results of the previous run already exist
in Airflow DB or filesystem (as a job files).
System submits error to DB if:
 - duplicate dag_id is found
 - duplicate job file is found in 'new'/'running' folder
 - job file hasn't been run at all and was moved to 'fail'
Refactored to run all DB related functions from one script.
Some bugs are fixed