Closes #4, display RNA-Seq bigWigs
For bedGraph the original SELECT is
used. For bigWig (not dUTP) _f_wtrack
is added. For bigWig dUTP, container
multiWig is added with upstream_f_wtrack
and downstream_f_wtrack. If both up-
and down-stream bigwig files are positive
and displayed with overlay, autoscale
works correct. If one of the bigWig file's
data is negative, autoscale works
independently for each of the bigWig file
so result is difficult to compare